all maps working
