+++ app/javascript/packs/application.js
@@ -1,8 +1,9 @@
 import "bootstrap";
 import 'mapbox-gl/dist/mapbox-gl.css'; // <-- you need to uncomment the stylesheet_pack_tag in the layout!
 import { initMapbox } from '../plugins/init_mapbox';
-import "../components/_goto_map";
+import { initGotoMap } from '../components/goto_map';
 
 import '@mapbox/mapbox-gl-geocoder/dist/mapbox-gl-geocoder.css';
 
 initMapbox();
+initGotoMap();
