+++ app/views/shared/_goto_map.html.erb
@@ -1,5 +1,5 @@
 <div
-  id="map"
+  id="goto-map"
   style="
   min-width: 100vw;
   min-height: 90vh;"
