+++ app/views/shared/_search_bar.html.erb
@@ -8,7 +8,7 @@
       placeholder: "Find a tour near you"
       %>
       <div class="form-inputs">
-        <%= button_to :submit %>
+        <%= button_to :search %>
       </div>
     <%end%>
   </div>
